Move TransferWindow
